Updated parsing scripts & files reflecting data from graphviz.org
Various layout changes on graphviz.org had broken the targeted parsing
in the existing scripts.

The scripts now depend slightly less on the specific HTML structure on
the page remaining constant. As long as we can find the right table
cell, it's sufficient to use $_->as_text() on that 'td' object to get
the string we are looking for, even if the target string is nested
inside multiple levels of tags (as it is on
https://graphviz.org/doc/info/attrs.html in July 2022).

In the resulting data files, the only change to the scripts' output was
in data/attributes.dat.

Leaving aside the reordering of some attributes in the file (which
reflects the latest order on the source page), the only material
changes in attributes.dat are:

* A new attribute 'nslimit1' is valid for graphs.
* Attribute 'tooltip' can now be used in all 4 contexts ('graph' was
  not a valid context before).